+++ projects/Mallard/src/components/article/html/__tests__/images.spec.ts
@@ -1,4 +1,4 @@
-import { renderCaption } from '../images'
+import { renderCaption } from '../components/images'
 
 describe('html', () => {
     describe('renderCaption', () => {
