+++ projects/Mallard/src/components/article/html/components/arrow.ts
