+++ projects/Mallard/src/components/article/html/components/icon/quotes.ts
