+++ projects/Mallard/src/components/article/html/components/pull-quote.ts
@@ -1,11 +1,12 @@
 import { css, html, px } from 'src/helpers/webview'
 import { PillarColours } from '@guardian/pasteup/palette'
-import { WrapLayout } from '../wrap/wrap'
+import { WrapLayout } from '../../wrap/wrap'
 import { families } from 'src/theme/typography'
 import { metrics } from 'src/theme/spacing'
 import { color } from 'src/theme/color'
 import { Breakpoints } from 'src/theme/breakpoints'
 import { Quotes } from './icon/quotes'
+import { CssProps } from '../helpers/css'
 
 const BubblePointer = () => html`
     <svg
@@ -28,13 +29,7 @@ const BubblePointer = () => html`
     </svg>
 `
 
-const quoteStyles = ({
-    colors,
-    wrapLayout,
-}: {
-    colors: PillarColours
-    wrapLayout: WrapLayout
-}) => css`
+const quoteStyles = ({ colors, wrapLayout }: CssProps) => css`
     blockquote {
         box-sizing: border-box;
         border: 1px solid ${colors.main};
