--- projects/Mallard/src/components/article/html/helpers/props.ts
@@ -1,7 +1,0 @@
-import { PillarColours } from '@guardian/pasteup/palette'
-import { WrapLayout } from '../../wrap/wrap'
-
-export interface CssProps {
-    colors: PillarColours
-    wrapLayout: WrapLayout
-}
