+++ projects/Mallard/src/components/article/types/article/webview.tsx
@@ -6,7 +6,7 @@ import { useArticle } from 'src/hooks/use-article'
 import { useImageSize } from 'src/hooks/use-image-size'
 import { useIssueSummary } from 'src/hooks/use-issue-summary'
 import { Article, PictureArticle } from '../../../../../../common/src'
-import { renderArticle } from '../../html/render'
+import { renderArticle } from '../../html/article'
 import { WrapLayout } from '../../wrap/wrap'
 import { ArticleTheme } from '../article'
 import { onShouldStartLoadWithRequest } from './helpers'
