+++ projects/Mallard/src/helpers/webview.ts
@@ -109,13 +109,11 @@ export const parsePing = (data: string) => JSON.parse(data) as WebViewPing
 export const makeHtml = ({
     styles,
     body,
-    rootProps,
 }: {
     styles: string
     body: string
-    rootProps?: string
 }) => html`
-    <html ${rootProps}>
+    <html>
         <head>
             <style>
                 ${styles}
